replace keytar, upgrade electron
